Delete unused code, and refactor
