+++ filesql_test.go
@@ -11,7 +11,6 @@ import (
 	"io"
 	"os"
 	"path/filepath"
-	"runtime"
 	"strconv"
 	"strings"
 	"sync"
@@ -4581,319 +4580,6 @@ func TestEdgeCasesMemoryLimits(t *testing.T) {
 		if err := rows.Err(); err != nil {
 			assert.NoError(t, err, "Rows error on wide file")
 			return
-		}
-	})
-}
-
-func TestMemoryLimitsAndLargeFiles(t *testing.T) {
-	t.Parallel()
-
-	t.Run("Large single cell content", func(t *testing.T) {
-		t.Parallel()
-
-		// Create a CSV with one very large cell (10MB)
-		const cellSize = 10 * 1024 * 1024 // 10MB
-		largeCellContent := strings.Repeat("x", cellSize)
-
-		content := fmt.Sprintf("col1,col2\n%s,value2\n", largeCellContent)
-
-		tmpFile, err := os.CreateTemp(t.TempDir(), "large_cell_*.csv")
-		if err != nil {
-			require.NoError(t, err, "Failed to create")
-		}
-		defer tmpFile.Close()
-
-		if _, err := tmpFile.WriteString(content); err != nil {
-			require.NoError(t, err, "Failed to write test content")
-		}
-
-		ctx, cancel := context.WithTimeout(context.Background(), 30*time.Second)
-		defer cancel()
-
-		db, err := OpenContext(ctx, tmpFile.Name())
-		assert.NoError(t, err, "Failed to handle large cell content")
-		defer db.Close()
-
-		// Verify we can query the data
-		tableName := strings.TrimSuffix(filepath.Base(tmpFile.Name()), ".csv")
-		tableName = strings.TrimSuffix(tableName, ".csv") // Handle double extensions
-
-		rows, err := db.QueryContext(ctx, fmt.Sprintf("SELECT LENGTH(col1) FROM \"%s\"", tableName))
-		assert.NoError(t, err, "Query failed on large cell")
-		defer rows.Close()
-		if err := rows.Err(); err != nil {
-			assert.NoError(t, err, "Rows error on large cell")
-			return
-		}
-
-		var length int
-		if rows.Next() {
-			if err := rows.Scan(&length); err != nil {
-				assert.NoError(t, err, "Scan failed")
-			} else if length != cellSize {
-				assert.Fail(t, "Expected cell length %d, got %d", cellSize, length)
-			}
-		}
-	})
-
-	t.Run("Many rows stress test", func(t *testing.T) {
-		t.Parallel()
-
-		// Skip this test in local development, only run on GitHub Actions
-		if os.Getenv("GITHUB_ACTIONS") == "" {
-			t.Skip("Skipping stress test in local development")
-		}
-
-		const numRows = 100000 // Full stress test on GitHub Actions
-
-		tmpFile, err := os.CreateTemp(t.TempDir(), "many_rows_*.csv")
-		if err != nil {
-			require.NoError(t, err, "Failed to create")
-		}
-		defer tmpFile.Close()
-
-		// Write header
-		if _, err := tmpFile.WriteString("id,name,value\n"); err != nil {
-			require.NoError(t, err, "Failed to write header")
-		}
-
-		// Write many rows
-		for i := range numRows {
-			line := fmt.Sprintf("%d,name_%d,value_%d\n", i, i, i)
-			if _, err := tmpFile.WriteString(line); err != nil {
-				require.NoError(t, err, "Failed to write row %d", i)
-			}
-		}
-
-		ctx, cancel := context.WithTimeout(context.Background(), 60*time.Second)
-		defer cancel()
-
-		db, err := OpenContext(ctx, tmpFile.Name())
-		assert.NoError(t, err, "Failed to handle many rows")
-		defer db.Close()
-
-		// Test aggregation query
-		tableName := strings.TrimSuffix(filepath.Base(tmpFile.Name()), ".csv")
-		tableName = strings.TrimSuffix(tableName, ".csv")
-
-		rows, err := db.QueryContext(ctx, fmt.Sprintf("SELECT COUNT(*) FROM \"%s\"", tableName))
-		assert.NoError(t, err, "Count query failed")
-		defer rows.Close()
-		if err := rows.Err(); err != nil {
-			assert.NoError(t, err, "Rows error on count query")
-			return
-		}
-
-		var count int
-		if rows.Next() {
-			if err := rows.Scan(&count); err != nil {
-				assert.NoError(t, err, "Count scan failed")
-			} else if count != numRows {
-				assert.Fail(t, "Expected %d rows, got %d", numRows, count)
-			}
-		}
-	})
-
-	t.Run("Memory usage monitoring", func(t *testing.T) {
-		t.Parallel()
-
-		// Record initial memory usage
-		var m1, m2 runtime.MemStats
-		runtime.GC()
-		runtime.ReadMemStats(&m1)
-
-		// Create multiple databases and close them
-		for i := range 10 {
-			content := fmt.Sprintf("col1,col2,col3\nvalue1_%d,value2_%d,value3_%d\n", i, i, i)
-
-			tmpFile, err := os.CreateTemp(t.TempDir(), fmt.Sprintf("memory_test_%d_*.csv", i))
-			if err != nil {
-				require.NoError(t, err, "Failed to create")
-			}
-
-			if _, err := tmpFile.WriteString(content); err != nil {
-				_ = tmpFile.Close() // Ignore error in test cleanup
-				require.NoError(t, err, "Failed to write content")
-			}
-			_ = tmpFile.Close() // Ignore error in test cleanup
-
-			ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-
-			db, err := OpenContext(ctx, tmpFile.Name())
-			if err != nil {
-				cancel()
-				assert.Fail(t, "Failed to open database %d: %v", i, err)
-				continue
-			}
-
-			// Perform some operations
-			tableName := strings.TrimSuffix(filepath.Base(tmpFile.Name()), ".csv")
-			tableName = strings.TrimSuffix(tableName, ".csv")
-
-			rows, err := db.QueryContext(ctx, fmt.Sprintf("SELECT * FROM \"%s\"", tableName))
-			if err != nil {
-				_ = db.Close() // Ignore error in test cleanup
-				cancel()
-				assert.Fail(t, "Query failed for database %d: %v", i, err)
-				continue
-			}
-			if err := rows.Err(); err != nil {
-				_ = rows.Close() // Ignore error in test cleanup
-				_ = db.Close()   // Ignore error in test cleanup
-				cancel()
-				assert.Fail(t, "Rows error for database %d: %v", i, err)
-				continue
-			}
-
-			for rows.Next() {
-				var col1, col2, col3 string
-				if err := rows.Scan(&col1, &col2, &col3); err != nil {
-					// Log error but continue test
-					t.Logf("Scan error: %v", err)
-				}
-			}
-			_ = rows.Close() // Ignore error in test cleanup
-			_ = db.Close()   // Ignore error in test cleanup
-			cancel()
-		}
-
-		// Check memory usage after cleanup
-		runtime.GC()
-		runtime.ReadMemStats(&m2)
-
-		memoryIncrease := int64(m2.Alloc) - int64(m1.Alloc) //nolint:gosec // Memory monitoring in test
-
-		// Set different thresholds based on runtime environment
-		// macOS tends to have higher memory usage due to different GC behavior
-		var threshold int64 = 50 * 1024 * 1024 // 50MB default
-		if runtime.GOOS == "darwin" {
-			threshold = 100 * 1024 * 1024 // 100MB for macOS
-		}
-
-		if memoryIncrease > threshold {
-			assert.LessOrEqual(t, memoryIncrease, threshold, "Memory usage increased by %d bytes (%.2f MB), may indicate memory leak (threshold: %.2f MB)",
-				memoryIncrease, float64(memoryIncrease)/(1024*1024), float64(threshold)/(1024*1024))
-		}
-	})
-
-	t.Run("Context cancellation during large file processing", func(t *testing.T) {
-		t.Parallel()
-
-		// Skip this test in local development, only run on GitHub Actions
-		if os.Getenv("GITHUB_ACTIONS") == "" {
-			t.Skip("Skipping large file processing test in local development")
-		}
-
-		// Create a moderately large file
-		const numRows = 10000 // Full test on GitHub Actions
-
-		tmpFile, err := os.CreateTemp(t.TempDir(), "cancellation_test_*.csv")
-		if err != nil {
-			require.NoError(t, err, "Failed to create")
-		}
-		defer tmpFile.Close()
-
-		// Write header and many rows
-		if _, err := tmpFile.WriteString("id,data\n"); err != nil {
-			require.NoError(t, err, "Failed to write header")
-		}
-
-		for i := range numRows {
-			line := fmt.Sprintf("%d,%s\n", i, strings.Repeat("data", 100))
-			if _, err := tmpFile.WriteString(line); err != nil {
-				require.NoError(t, err, "Failed to write row %d", i)
-			}
-		}
-
-		// Create context with very short timeout to trigger cancellation
-		ctx, cancel := context.WithTimeout(context.Background(), 1*time.Millisecond)
-		defer cancel()
-
-		// This should either succeed quickly or fail due to timeout
-		db, err := OpenContext(ctx, tmpFile.Name())
-
-		if err != nil {
-			// Timeout error is acceptable
-			if ctx.Err() == context.DeadlineExceeded {
-				return // Expected behavior
-			}
-			// Other errors might indicate problems with context handling
-			t.Logf("Context cancellation test completed with error (may be expected): %v", err)
-		}
-
-		if db != nil {
-			_ = db.Close() // Ignore error in test cleanup
-		}
-	})
-
-	t.Run("Chunk size configuration test", func(t *testing.T) {
-		t.Parallel()
-
-		// Create a file that's larger than default chunk size
-		const numRows = 1000
-		content := "col1,col2,col3\n"
-		for i := range numRows {
-			content += fmt.Sprintf("value_%d,%s,data_%d\n", i, strings.Repeat("x", 1000), i)
-		}
-
-		tmpFile, err := os.CreateTemp(t.TempDir(), "chunk_test_*.csv")
-		if err != nil {
-			require.NoError(t, err, "Failed to create")
-		}
-		defer tmpFile.Close()
-
-		if _, err := tmpFile.WriteString(content); err != nil {
-			require.NoError(t, err, "Failed to write content")
-		}
-
-		ctx := context.Background()
-
-		// Test with different chunk sizes
-		chunkSizes := []int{1024, 10240, 102400} // 1KB, 10KB, 100KB
-
-		for _, chunkSize := range chunkSizes {
-			t.Run(fmt.Sprintf("ChunkSize_%d", chunkSize), func(t *testing.T) {
-				builder := NewBuilder().
-					AddPath(tmpFile.Name()).
-					SetDefaultChunkSize(chunkSize)
-
-				validatedBuilder, err := builder.Build(ctx)
-				if err != nil {
-					assert.Fail(t, "Build failed with chunk size %d: %v", chunkSize, err)
-					return
-				}
-
-				db, err := validatedBuilder.Open(ctx)
-				if err != nil {
-					assert.Fail(t, "Open failed with chunk size %d: %v", chunkSize, err)
-					return
-				}
-				defer db.Close()
-
-				// Verify data integrity regardless of chunk size
-				tableName := strings.TrimSuffix(filepath.Base(tmpFile.Name()), ".csv")
-				tableName = strings.TrimSuffix(tableName, ".csv")
-
-				rows, err := db.QueryContext(ctx, fmt.Sprintf("SELECT COUNT(*) FROM \"%s\"", tableName))
-				if err != nil {
-					assert.Fail(t, "Count query failed with chunk size %d: %v", chunkSize, err)
-					return
-				}
-				defer rows.Close()
-				if err := rows.Err(); err != nil {
-					assert.Fail(t, "Rows error with chunk size %d: %v", chunkSize, err)
-					return
-				}
-
-				var count int
-				if rows.Next() {
-					if err := rows.Scan(&count); err != nil {
-						assert.Fail(t, "Count scan failed with chunk size %d: %v", chunkSize, err)
-					} else if count != numRows {
-						assert.Fail(t, "Expected %d rows with chunk size %d, got %d", numRows, chunkSize, count)
-					}
-				}
-			})
 		}
 	})
 }
